# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateStructureRequest:
    properties:
      active:
        example: true
        type: boolean
      address:
        example: Jakarta
        type: string
      customer_id:
        type: integer
      level:
        example: 1
        type: integer
      name:
        example: Board of Directors
        type: string
      parentKey:
        example: "null"
        type: string
      tempKey:
        example: "1"
        type: string
    required:
    - customer_id
    - level
    - name
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
  entity.Activity:
    properties:
      activity_checkins:
        items:
          $ref: '#/definitions/entity.ActivityCheckin'
        type: array
      agenda:
        type: string
      attendees:
        items:
          $ref: '#/definitions/entity.User'
        type: array
      created_at:
        type: string
      created_by:
        type: integer
      creator:
        $ref: '#/definitions/entity.User'
      customer_id:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      location_name:
        type: string
      start_time:
        type: string
      status:
        type: string
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  entity.ActivityCheckin:
    properties:
      activity_id:
        type: integer
      checked_in_at:
        type: string
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
  entity.Address:
    properties:
      active:
        type: boolean
      address:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      customer_id:
        type: integer
      id:
        type: integer
      main:
        type: boolean
      name:
        description: SupplierID *uint          `json:"supplier_id"` // HAPUS field
          ini
        type: string
      postal_code:
        type: string
      state:
        type: string
      street:
        type: string
      updated_at:
        type: string
    type: object
  entity.Contact:
    properties:
      active:
        type: boolean
      birthdate:
        type: string
      created_at:
        type: string
      customer_id:
        type: integer
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      job_position:
        type: string
      main:
        type: boolean
      mobile:
        type: string
      name:
        type: string
      phone:
        type: string
      position:
        type: string
      updated_at:
        type: string
    type: object
  entity.Customer:
    properties:
      account_manager_id:
        type: string
      activities:
        items:
          $ref: '#/definitions/entity.Activity'
        type: array
      addresses:
        description: Relations
        items:
          $ref: '#/definitions/entity.Address'
        type: array
      average_cost:
        type: number
      brand_name:
        type: string
      category:
        type: string
      code:
        type: string
      contacts:
        items:
          $ref: '#/definitions/entity.Contact'
        type: array
      created_at:
        type: string
      description:
        type: string
      email:
        type: string
      groups:
        items:
          $ref: '#/definitions/entity.Group'
        type: array
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
      others:
        items:
          $ref: '#/definitions/entity.Other'
        type: array
      phone:
        type: string
      rating:
        type: number
      sosmeds:
        items:
          $ref: '#/definitions/entity.Sosmed'
        type: array
      status:
        description: Status internal
        type: string
      structures:
        items:
          $ref: '#/definitions/entity.Structure'
        type: array
      updated_at:
        type: string
      website:
        type: string
    type: object
  entity.Group:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      customers:
        description: Relations
        items:
          $ref: '#/definitions/entity.Customer'
        type: array
      id:
        type: integer
      name_group:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  entity.Other:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      customer_id:
        type: integer
      id:
        type: integer
      key:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  entity.Role:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role_name:
        type: string
      updated_at:
        type: string
      users:
        description: Relations
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
  entity.Sosmed:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      customer_id:
        type: integer
      followers:
        type: integer
      handle:
        type: string
      id:
        type: integer
      name:
        type: string
      platform:
        type: string
      updated_at:
        type: string
      url:
        type: string
      username:
        type: string
    type: object
  entity.Structure:
    properties:
      active:
        type: boolean
      address:
        type: string
      children:
        items:
          $ref: '#/definitions/entity.Structure'
        type: array
      created_at:
        type: string
      customer_id:
        type: integer
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      parent:
        $ref: '#/definitions/entity.Structure'
      parent_id:
        type: integer
      position:
        type: integer
      updated_at:
        type: string
    type: object
  entity.User:
    properties:
      activities:
        items:
          $ref: '#/definitions/entity.Activity'
        type: array
      activity_checkins:
        items:
          $ref: '#/definitions/entity.ActivityCheckin'
        type: array
      attending_activities:
        items:
          $ref: '#/definitions/entity.Activity'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/entity.Role'
        description: |-
          Update struct User, tambahkan di bagian Relations:
          Relations
      role_id:
        description: Default to regular user role
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: API untuk manajemen customer menggunakan Golang, Gin, dan PostgreSQL
  title: Customer Management API
  version: "1.0"
paths:
...
  /api/customers/{id}/full:
    get:
      consumes:
      - application/json
      description: Get complete customer data with all relations
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Customer'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get customer full data
      tags:
      - Customers
...
  /api/customers/{id}/structures:
    get:
      consumes:
      - application/json
      description: Get all organizational structures for specific customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Structure'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get customer structures
      tags:
      - Structures
    post:
      consumes:
      - application/json
      - application/json
      description: |-
        Create a new organizational structure for specific customer
        Create a new structure
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Structure data
        in: body
        name: structure
        required: true
        schema:
          $ref: '#/definitions/entity.Structure'
      - description: Structure data
        in: body
        name: structure
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStructureRequest'
      produces:
      - application/json
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Structure'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Create structure
      tags:
      - Structures
      - Structures
  /api/customers/{id}/structures/by-level:
    get:
      consumes:
      - application/json
      description: Get organizational structures filtered by level for specific customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Structure level filter
        in: query
        name: level
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Structure'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get structures by level
      tags:
      - Structures
...
  /api/customers/{id}/with-all:
    get:
      consumes:
      - application/json
      description: Get customer data with all related data (addresses, sosmeds, contacts,
        structures)
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Customer'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get customer with all relations
      tags:
      - Customers
...
  /api/customers/{id}/with-structures:
    get:
      consumes:
      - application/json
      description: Get customer data with all organizational structures
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Customer'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get customer with structures
      tags:
      - Customers
...
  /api/structures:
    post:
      consumes:
      - application/json
      - application/json
      description: |-
        Create a new organizational structure for specific customer
        Create a new structure
      parameters:
      - description: Structure data
        in: body
        name: structure
        required: true
        schema:
          $ref: '#/definitions/entity.Structure'
      - description: Structure data
        in: body
        name: structure
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStructureRequest'
      produces:
      - application/json
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Structure'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Create structure
      tags:
      - Structures
      - Structures
  /api/structures/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an organizational structure by ID
      parameters:
      - description: Structure ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete structure
      tags:
      - Structures
    get:
      consumes:
      - application/json
      description: Get a specific organizational structure by ID
      parameters:
      - description: Structure ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Structure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get structure by ID
      tags:
      - Structures
    put:
      consumes:
      - application/json
      description: Update an existing organizational structure
      parameters:
      - description: Structure ID
        in: path
        name: id
        required: true
        type: integer
      - description: Structure data
        in: body
        name: structure
        required: true
        schema:
          $ref: '#/definitions/entity.Structure'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Structure'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update structure
      tags:
      - Structures
...
securityDefinitions:
  BearerAuth:
    description: Type \"Bearer\" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"